# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Listings Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Вывод всех бронирований пользователя",
                "operationId": "39cda3cc485672c7b4532fac2e444926",
                "responses": {
                    "200": {
                        "description": "Список бронирований"
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Создание бронирования",
                "operationId": "76856b1825fd8f8255a04a1619d6495e",
                "parameters": [
                    {
                        "name": "listing_id",
                        "in": "query",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "check_in",
                        "in": "query",
                        "description": "Дата въезда",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "check_out",
                        "in": "query",
                        "description": "Дата выезда",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование создано"
                    },
                    "404": {
                        "description": "Пользователь или объявление не найдены"
                    },
                    "400": {
                        "description": "Даты недоступны для бронирования"
                    }
                }
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Вывод информации о бронировании",
                "operationId": "fe233850da5548f6e72cf5e709e2e28f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование"
                    },
                    "404": {
                        "description": "Бронирование не найдено"
                    }
                }
            },
            "put": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Изменение бронирования",
                "operationId": "7d83fb8e0b692333f812e27411aa3f1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "check_in",
                        "in": "query",
                        "description": "Новая дата въезда",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "check_out",
                        "in": "query",
                        "description": "Новая дата выезда",
                        "required": false,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Новый статус бронирования",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование изменено"
                    },
                    "404": {
                        "description": "Бронирование не найдено"
                    },
                    "403": {
                        "description": "Вы не владелец бронирования"
                    },
                    "400": {
                        "description": "Даты недоступны для бронирования"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Удаление бронирования",
                "operationId": "37664c0aa2659691af768e7b7c8b6402",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронирование удалено"
                    },
                    "404": {
                        "description": "Бронирование не найдено"
                    },
                    "403": {
                        "description": "Вы не владелец бронирования"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}